+++ apps/ledger-live-desktop/src/renderer/modals/NoFundsStake/index.tsx
