--- apps/ledger-live-desktop/src/renderer/modals/Stake/types.ts
@@ -1,4 +1,0 @@
-export interface IconProps {
-  size: number;
-  color: string;
-}
