+++ apps/ledger-live-desktop/src/renderer/modals/index.js
@@ -12,7 +12,7 @@ import MODAL_RECEIVE from "./Receive";
 import MODAL_SEND from "./Send";
 import MODAL_SIGN_MESSAGE from "./SignMessage";
 import MODAL_SIGN_TRANSACTION from "./SignTransaction";
-import MODAL_NO_FUNDS_STAKE from "./Stake";
+import MODAL_NO_FUNDS_STAKE from "./NoFundsStake";
 import MODAL_UPDATE_FIRMWARE from "./UpdateFirmwareModal";
 import MODAL_MIGRATE_ACCOUNTS from "./MigrateAccounts";
 import MODAL_EXPORT_ACCOUNTS from "./ExportAccounts";
